+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Layout.kt
@@ -46,7 +46,6 @@ internal fun Layout(
 		measurePolicy = NoContentMeasurePolicyMeasurePolicy(measurePolicy),
 		debugPolicy = { debugInfo() + " x=$x y=$y w=$width h=$height${modifier.toDebugString()}" },
 		modifier = modifier,
-		factory = NodeFactory,
 	)
 }
 
@@ -83,7 +82,6 @@ public fun Layout(
 		},
 		modifier = modifier,
 		content = content,
-		factory = NodeFactory,
 	)
 }